PR feedback: move new test into file system test
