+++ test/FileSystemInfo.unittest.js
@@ -5,6 +5,11 @@ const util = require("util");
 const FileSystemInfo = require("../lib/FileSystemInfo");
 const { buffersSerializer } = require("../lib/util/serialization");
 
+const fs = require("fs");
+const path = require("path");
+const { lstatReadlinkAbsolute } = require("../lib/util/fs");
+jest.mock("fs");
+
 describe("FileSystemInfo", () => {
 	const files = [
 		"/path/file.txt",
@@ -432,5 +437,41 @@ ${details(snapshot)}`)
 				done();
 			});
 		});
+	});
+});
+
+describe("lstatReadlinkAbsolute", () => {
+	it("should call the callback with an error if fs.readlink fails", done => {
+		// We will mock fs.readlink to always return this error
+		const mockError = new Error("Mocked error");
+
+		// Mock fs.lstat to always return a symbolic link
+		const fsLstat = jest
+			.spyOn(fs, "lstat")
+			.mockImplementation((_, callback) => {
+				callback(null, { isSymbolicLink: () => true });
+			});
+
+		// Mock fs.readlink to always return the error above
+		const fsReadlink = jest
+			.spyOn(fs, "readlink")
+			.mockImplementation((_, callback) => {
+				callback(mockError);
+			});
+
+		const callback = (err, result) => {
+			try {
+				expect(err).toBe(mockError);
+				expect(result).toBeUndefined();
+				done();
+			} catch (err_) {
+				done(err_);
+			}
+		};
+
+		lstatReadlinkAbsolute(fs, path.resolve("/some/path"), callback);
+
+		fsLstat.mockReset();
+		fsReadlink.mockReset();
 	});
 });
